+++ ms-quartos/src/main/java/br/com/fiap/hackathon/quartos/entity/Quarto.java
@@ -15,6 +15,6 @@ public class Quarto {
   private TipoQuarto tipo;
   private int totalPessoas;
   private int totalCamas;
-  private BigDecimal valorDiaria;
+  private double valorDiaria;
   private List<Movel> moveis;
 }
